requirements.txt
Flask==3.1.0
Werkzeug==3.1.3
flask-cors==4.0.0
python-dotenv==1.0.0
pymongo==4.11.2
PyPDF2==3.0.1
python-docx==0.8.11
Pillow==11.0.0
